add client & update-codegen.sh
