+++ PROJECT
@@ -7,19 +7,25 @@ resources:
   controller: true
   domain: streamnative.io
   kind: FunctionMesh
-  path: github.com/streamnative/function-mesh/api/v1alpha1
+  path: github.com/streamnative/function-mesh/api/compute/v1alpha1
   version: v1alpha1
 - group: compute
   controller: true
   domain: streamnative.io
   kind: Source
-  path: github.com/streamnative/function-mesh/api/v1alpha1
+  path: github.com/streamnative/function-mesh/api/compute/v1alpha1
   version: v1alpha1
 - group: compute
   controller: true
   domain: streamnative.io
   kind: Sink
-  path: github.com/streamnative/function-mesh/api/v1alpha1
+  path: github.com/streamnative/function-mesh/api/compute/v1alpha1
+  version: v1alpha1
+- group: compute
+  controller: true
+  domain: streamnative.io
+  kind: Function
+  path: github.com/streamnative/function-mesh/api/compute/v1alpha1
   version: v1alpha1
 version: "3"
 plugins:
