+++ api/compute/v1alpha1/function_types.go
@@ -103,6 +103,7 @@ type FunctionStatus struct {
 	ObservedGeneration int64                           `json:"observedGeneration,omitempty"`
 }
 
+// +genclient
 // +kubebuilder:object:root=true
 // +kubebuilder:subresource:status
 //+kubebuilder:subresource:scale:specpath=.spec.replicas,statuspath=.status.replicas,selectorpath=.status.selector
