+++ api/compute/v1alpha1/functionmesh_types.go
@@ -44,6 +44,7 @@ type FunctionMeshStatus struct {
 	ObservedGeneration int64                        `json:"observedGeneration,omitempty"`
 }
 
+// +genclient
 // +kubebuilder:object:root=true
 // +kubebuilder:subresource:status
