+++ controllers/sink_controller_test.go
@@ -31,7 +31,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/streamnative/function-mesh/api/v1alpha1"
+	"github.com/streamnative/function-mesh/api/compute/v1alpha1"
 )
 
 var _ = Describe("Sink Controller", func() {
