+++ controllers/spec/common_test.go
@@ -21,13 +21,10 @@ import (
 	"strings"
 	"testing"
 
-	corev1 "k8s.io/api/core/v1"
-
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	"github.com/streamnative/function-mesh/api/v1alpha1"
-
+	"github.com/streamnative/function-mesh/api/compute/v1alpha1"
 	"github.com/stretchr/testify/assert"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func TestGetDownloadCommand(t *testing.T) {
