+++ controllers/spec/function_mesh.go
@@ -18,7 +18,7 @@
 package spec
 
 import (
-	"github.com/streamnative/function-mesh/api/v1alpha1"
+	"github.com/streamnative/function-mesh/api/compute/v1alpha1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
