+++ controllers/spec/source.go
@@ -19,7 +19,7 @@ package spec
 
 import (
 	"github.com/gogo/protobuf/jsonpb"
-	"github.com/streamnative/function-mesh/api/v1alpha1"
+	"github.com/streamnative/function-mesh/api/compute/v1alpha1"
 	appsv1 "k8s.io/api/apps/v1"
 	autov2beta2 "k8s.io/api/autoscaling/v2beta2"
 	corev1 "k8s.io/api/core/v1"
