+++ controllers/spec/utils.go
@@ -23,12 +23,13 @@ import (
 	"regexp"
 	"strings"
 
+	v1alpha1 "github.com/streamnative/function-mesh/api/compute/v1alpha1"
+
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
 
 	"k8s.io/apimachinery/pkg/util/validation"
 
-	"github.com/streamnative/function-mesh/api/v1alpha1"
 	"github.com/streamnative/function-mesh/controllers/proto"
 )
