+++ controllers/spec/utils_test.go
@@ -20,7 +20,7 @@ package spec
 import (
 	"testing"
 
-	"github.com/streamnative/function-mesh/api/v1alpha1"
+	"github.com/streamnative/function-mesh/api/compute/v1alpha1"
 
 	"github.com/stretchr/testify/assert"
 )
