+++ tools/migration.go
@@ -24,7 +24,7 @@ import (
 	"strings"
 
 	"github.com/ghodss/yaml"
-	"github.com/streamnative/function-mesh/api/v1alpha1"
+	"github.com/streamnative/function-mesh/api/compute/v1alpha1"
 	cmdutils "github.com/streamnative/pulsarctl/pkg/cmdutils"
 	"github.com/streamnative/pulsarctl/pkg/pulsar/common"
 	corev1 "k8s.io/api/core/v1"
